predict name identifiers
