+++ Content.Client/NameIdentifier/NameIdentifierSystem.cs
@@ -1,0 +1,5 @@
+using Content.Shared.NameIdentifier;
+
+namespace Content.Client.NameIdentifier;
+
+public sealed class NameIdentifierSystem : SharedNameIdentifierSystem;
